+++ packages/metro-config/src/assetPluginForMonorepos.js
@@ -1,6 +1,11 @@
 /**
  * @typedef {import("metro").AssetData} AssetData;
+ * @typedef {import("metro-config").ConfigT} ConfigT;
  * @typedef {import("metro-config").Middleware} Middleware;
+ *
+ * @typedef {import("metro/src/Server") & {
+ *   _config?: ConfigT;
+ * }} Server;
  */
 
 /**
@@ -25,12 +30,39 @@ function assetPlugin(assetData) {
 }
 
 /**
+ * Injects {@link assetPlugin} into server asset plugins.
+ *
+ * This should only be called from {@link enhanceMiddleware} to ensure that the
+ * asset plugin is only applied when we are serving. This has the nice
+ * side-effect that the plugin doesn't get included if the middleware is
+ * unused (during bundling) or removed.
+ *
+ * @param {Server} server
+ */
+function injectAssetPlugin(server) {
+  const config = server._config;
+  if (!config || !Array.isArray(config.transformer.assetPlugins)) {
+    console.warn(
+      "'@rnx-kit/metro-config' was unable to install the asset plugin for " +
+        "monorepos. Please try again with the latest version. If this " +
+        "warning still persists, you can file an issue at " +
+        "https://github.com/microsoft/rnx-kit/issues/new?assignees=&labels=bug&template=bug_report.yml"
+    );
+    return;
+  }
+
+  config.transformer.assetPlugins.push(__filename);
+}
+
+/**
  * This middleware restores `..` in asset URLs.
  *
  * @param {Middleware} middleware
+ * @param {Server} server
  * @returns {Middleware}
  */
-function enhanceMiddleware(middleware) {
+function enhanceMiddleware(middleware, server) {
+  injectAssetPlugin(server);
   return (req, res, next) => {
     const { url } = req;
     if (url && url.startsWith("/assets/")) {