+++ packages/metro-config/test/index.test.ts
@@ -1,5 +1,6 @@
 import * as fs from "node:fs";
 import * as path from "node:path";
+import { enhanceMiddleware } from "../src/assetPluginForMonorepos";
 import {
   defaultWatchFolders,
   exclusionList,
@@ -9,16 +10,6 @@ import {
 } from "../src/index";
 
 describe("@rnx-kit/metro-config", () => {
-  const metroConfigKeys = [
-    "cacheStores",
-    "resolver",
-    "serializer",
-    "server",
-    "symbolicator",
-    "transformer",
-    "watchFolders",
-  ];
-
   const currentWorkingDir = process.cwd();
 
   /**
@@ -239,8 +230,26 @@ describe("@rnx-kit/metro-config", () => {
     ).toBeTruthy();
     expect(conanExclude.test("Test.ProjectImports.zip")).toBeTruthy();
   });
+});
 
-  test("makeMetroConfig() returns a default Metro config", async () => {
+describe("makeMetroConfig", () => {
+  const consoleWarnSpy = jest.spyOn(require("@rnx-kit/console"), "warn");
+
+  const metroConfigKeys = [
+    "cacheStores",
+    "resolver",
+    "serializer",
+    "server",
+    "symbolicator",
+    "transformer",
+    "watchFolders",
+  ];
+
+  afterEach(() => {
+    consoleWarnSpy.mockReset();
+  });
+
+  test("returns a default Metro config", async () => {
     const config = makeMetroConfig();
     expect(Object.keys(config).sort()).toEqual(metroConfigKeys);
 
@@ -252,6 +261,8 @@ describe("@rnx-kit/metro-config", () => {
       fail("Expected `config.resolver.blacklistRE` to be a RegExp");
     } else if (!(config.resolver.blockList instanceof RegExp)) {
       fail("Expected `config.resolver.blockList` to be a RegExp");
+    } else if (!config.server) {
+      fail("Expected `config.server` to be defined");
     } else if (!config.transformer) {
       fail("Expected `config.transformer` to be defined");
     } else if (!config.transformer.getTransformOptions) {
@@ -268,6 +279,9 @@ describe("@rnx-kit/metro-config", () => {
     expect(config.resolver.blacklistRE.source).toBe(blockList);
     expect(config.resolver.blockList.source).toBe(blockList);
 
+    expect(config.server.enhanceMiddleware).toBe(enhanceMiddleware);
+    expect(config.transformer.assetPlugins).toBeUndefined();
+
     const opts = { dev: false, hot: false };
     const transformerOptions = await config.transformer.getTransformOptions(
       [],
@@ -282,7 +296,7 @@ describe("@rnx-kit/metro-config", () => {
     expect(config.watchFolders.length).toBeGreaterThan(0);
   });
 
-  test("makeMetroConfig() merges Metro configs", async () => {
+  test("merges Metro configs", async () => {
     const config = makeMetroConfig({
       projectRoot: __dirname,
       resetCache: true,
@@ -303,6 +317,8 @@ describe("@rnx-kit/metro-config", () => {
       fail("Expected `config.resolver.blacklistRE` to be a RegExp");
     } else if (!(config.resolver.blockList instanceof RegExp)) {
       fail("Expected `config.resolver.blockList` to be a RegExp");
+    } else if (!config.server) {
+      fail("Expected `config.server` to be defined");
     } else if (!config.transformer) {
       fail("Expected `config.transformer` to be defined");
     } else if (!config.transformer.getTransformOptions) {
@@ -319,6 +335,9 @@ describe("@rnx-kit/metro-config", () => {
     expect(config.resolver.blacklistRE.source).toBe(blockList);
     expect(config.resolver.blockList.source).toBe(blockList);
 
+    expect(config.server.enhanceMiddleware).toBe(enhanceMiddleware);
+    expect(config.transformer.assetPlugins).toBeUndefined();
+
     const opts = { dev: false, hot: false };
     const transformerOptions = await config.transformer.getTransformOptions(
       [],
@@ -333,7 +352,7 @@ describe("@rnx-kit/metro-config", () => {
     expect(config.watchFolders.length).toBeGreaterThan(0);
   });
 
-  test("makeMetroConfig() merges `extraNodeModules`", () => {
+  test("merges `extraNodeModules`", () => {
     const config = makeMetroConfig({
       projectRoot: __dirname,
       resolver: {
@@ -359,7 +378,7 @@ describe("@rnx-kit/metro-config", () => {
     expect(extraNodeModules["react-native"]).toBe("/skynet");
   });
 
-  test("makeMetroConfig() sets both `blacklistRE` and `blockList`", () => {
+  test("sets both `blacklistRE` and `blockList`", () => {
     const configWithBlacklist = makeMetroConfig({
       resolver: {
         blacklistRE: /.*/,