+++ packages/metro-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log - @rnx-kit/metro-config
 
+## 1.2.39
+
+### Patch Changes
+
+- 3e706f3e: Fix monorepo asset plugin breaking bundling
+
+## 1.2.38
+
+### Patch Changes
+
+- a9296175: Verify `server.enhanceMiddleware` is only set if `transformer.assetPlugins` includes the asset plugin.
+
 ## 1.2.37
 
 ### Patch Changes